pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "personamem0"
version = "0.1.0"
description = "PersonaMem benchmark and Mem0 experiment tooling"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "mem0ai",
  "openai",
  "python-dotenv",
  "tqdm",
  "matplotlib",
  "pandas",
  "datasets",
  "huggingface_hub",
]

[project.optional-dependencies]
dev = [
  "ipykernel",
  "jupyter",
]

[project.scripts]
personamem-benchmark = "personamem.personamem_benchmark:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
